Update App.js
Add radio boxes to form
Add employment empty string to state to be used for radio boxes
Add radio boxes by adding fieldset, legend, and 3 radio inputs
In order for React to manage the checked property of the radio box, we have to add the checked property with formData.employment and make it a boolean value by making it equal to the value of the input. In the case if the employment field equals to unemployed, for example, then, the checked box should be marked.
The label tag is to identify the check box radio, and the htmlFor makes it so that when clicking on the text, the checkbox is clicked as well.